:recycle: refactor: 디렉토리 구조 변경
디렉 토리 구조 변경으로 인한 위치 수정

Fixes: #96
