+++ src/main/java/com/example/umcmatchingcenter/service/myProjectService/MyProjectService.java
@@ -18,7 +18,7 @@ import com.example.umcmatchingcenter.dto.ProjectDTO.PartMatchingResponseDTO;
 import com.example.umcmatchingcenter.dto.ProjectDTO.TotalMatchingResponseDTO;
 import com.example.umcmatchingcenter.service.AlarmService.AlarmCommandService;
 import com.example.umcmatchingcenter.service.memberService.MemberQueryService;
-import com.example.umcmatchingcenter.service.queryService.RecruitmentQueryService;
+import com.example.umcmatchingcenter.service.recruitmentService.RecruitmentQueryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
